pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "playwright-python-framework"
version = "0.1.0"
description = "Playwright + Pytest automation framework scaffold"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "playwright==1.52.0",
    "pytest==8.3.5",
    "pytest-xdist==3.8.0",
    "pytest-playwright==0.7.0",
    "python-dotenv==1.1.0",
    "ruff==0.11.8",
    "pre-commit==4.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
    "test_case_id(id): Link a test with a test case identifier",
    "key(id): External test management key mapping",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.setuptools.packages.find]
include = ["actions*", "helpers*", "page_objects*", "tests*"]
